Include 'email_verified', + all email data in extra response
This allows consumers to check whether or not Github has verified the
emails.

Manual selection by consumers is also available using the emails array
in the `extra` response data.

Additionally, a new 'display_email' field is added to `info`, which will
always be the email displayed on your GitHub user page.